# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/image/remove/{userID}/{petID}": {
            "delete": {
                "description": "Removes an image for a specific pet associated with a user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image"
                ],
                "summary": "Removes an image for a specific pet associated with a user.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pet ID",
                        "name": "petID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Image removed successfully",
                        "schema": {
                            "$ref": "#/definitions/model.RemoveImageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid input data",
                        "schema": {
                            "$ref": "#/definitions/model.RemoveImageResponse"
                        }
                    },
                    "404": {
                        "description": "Image not found",
                        "schema": {
                            "$ref": "#/definitions/model.RemoveImageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.RemoveImageResponse"
                        }
                    }
                }
            }
        },
        "/api/image/upload/{userID}/{petID}": {
            "post": {
                "description": "Uploads an image for a specific pet associated with a user.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image"
                ],
                "summary": "Uploads an image for a specific pet associated with a user.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pet ID",
                        "name": "petID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image file",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image uploaded successfully",
                        "schema": {
                            "$ref": "#/definitions/model.UploadImageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid input data",
                        "schema": {
                            "$ref": "#/definitions/model.UploadImageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.UploadImageResponse"
                        }
                    }
                }
            }
        },
        "/api/pet/create": {
            "post": {
                "description": "Create a new pet with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Create a new pet",
                "parameters": [
                    {
                        "description": "Pet object in JSON format",
                        "name": "pet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Pet"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created pet",
                        "schema": {
                            "$ref": "#/definitions/model.Pet"
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "description": "Register a new user with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User object in JSON format",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully registered user",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Pet": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "breed_id": {
                    "type": "integer"
                },
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "food": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lives_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "passport_id": {
                    "type": "integer"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "spicies": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "todo_id": {
                    "type": "integer"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "model.RemoveImageResponse": {
            "type": "object"
        },
        "model.UploadImageResponse": {
            "type": "object",
            "properties": {
                "imageURL": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "repeat_password": {
                    "type": "string"
                }
            }
        }
    }
}